+++ nvim/config/lua/user/options.lua
