+++ nvim/config/lua/user/trouble.lua
